moving out multiprocessing routine and commenting
