+++ cea/resources/radiation_daysim/default_materials.rad
@@ -8,37 +8,22 @@ void plastic wallT5
 0
 5 0.5 0.5 0.5 0.03 0.02
 
-void glass winT1
-0
-0
-3 0.96 0.96 0.96
-
-void plastic roofT2
-0
-0
-5 0.5 0.5 0.5 0.03 0.02
-
-void plastic wallT2
-0
-0
-5 0.5 0.5 0.5 0.03 0.02
-
-void plastic wallT4
-0
-0
-5 0.5 0.5 0.5 0.03 0.02
-
-void plastic roofT1
-0
-0
-5 0.5 0.5 0.5 0.03 0.02
-
 void glass winT2
 0
 0
 3 0.96 0.96 0.96
 
 void plastic roofT4
+0
+0
+5 0.5 0.5 0.5 0.03 0.02
+
+void glass winT1
+0
+0
+3 0.96 0.96 0.96
+
+void plastic roofT1
 0
 0
 5 0.5 0.5 0.5 0.03 0.02
